src/backend/requirements.txt
aioredis==1.3.1
appdirs==1.4.4
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.12.3
Automat==20.2.0
black==20.8b1
cffi==1.14.4
channels==3.0.3
channels-redis==3.2.0
click==7.1.2
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.1
Django==3.1.4
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
hiredis==1.1.0
hyperlink==20.0.1
idna==3.0
incremental==17.5.0
msgpack==1.0.2
mypy-extensions==0.4.3
pathspec==0.8.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==2.0.0
pyOpenSSL==20.0.1
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
regex==2020.11.13
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
Twisted==20.3.0
txaio==20.12.1
typed-ast==1.4.2
typing-extensions==3.7.4.3
zope.interface==5.2.0
